# DateNDash THE APP IDEA
# DateNDash THE APP IDEA
anonymously chat with ppl your near and only go public to that person...But.. @3am you and all info gets DELETED..each time you use the app.. your generated a user name and time stamp.. app does not give precise location, only neighborhood proximity. its like speed dating on the go...
also have the choice of deleting all info in a short time that user selects. but.. the main function of the app is to SPEED DATE BASED ON PROXIMITY.
with the choice of being anonymouse or not.